+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+### Added
+- Status of various part HEAppE and way to obtain error logs
+
+### New Endpoints:
+
+POST /heappe/Management/Status
+POST /heappe/Management/StatusErrorLogs
+
 ## V6.0.0
 
 ### Changed